% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{areanorm}
\alias{areanorm}
\title{Normalization of spectra}
\usage{
areanorm(spectra)
}
\arguments{
\item{spectra}{matrix of vector with spectral data}
}
\description{
normalize spectra to unit area
}

