% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{getColors}
\alias{getColors}
\title{Returns colors from colorbrewer2 palette}
\usage{
getColors(ncols = 1)
}
\arguments{
\item{ncols}{number of colors}
}
\description{
Returns colors from colorbrewer2 palette
}

