% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/purity.R
\name{getpurevars}
\alias{getpurevars}
\title{Identifies pure variables}
\usage{
getpurevars(D, ncomp, offset = 5, exclude = NULL)
}
\arguments{
\item{D}{matrix with the spectra or their inverted second derivative}

\item{ncomp}{number of pure components}

\item{offset}{offset in percent for calculation of parameter alpha}

\item{exclude}{which columns exclude when detecting the pure variables}
}
\value{
The function returns a list with with following fields:
\item{ncomp }{number of pure components.}
\item{purvars }{vector with indices for pure variables.}
\item{purityspec }{matrix with purity values for each resolved components.}
\item{stdspec }{matrix with weighted standard deviation values.}
\item{purity }{vector with purity values for resolved components.}
}
\description{
The method identifies indices of pure variables using the SIMPLISMA
algorithm.
}

