% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mcrals.R
\name{image.mcrals}
\alias{image.mcrals}
\title{Image with concentration map}
\usage{
\method{image}{mcrals}(x, ...)
}
\arguments{
\item{x}{object with mcrals results}

\item{...}{other parameters}
}
\description{
Make an image with map of resolved concentrations
}
\details{
The method is similar to \code{\link{image.purity}}.
}

