% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/purity.R
\name{image.purity}
\alias{image.purity}
\title{Image with concentration map}
\usage{
\method{image}{purity}(x, ncomp, main = NULL, col = getColors(256),
  xaxt = "n", yaxt = "n", ...)
}
\arguments{
\item{x}{object with purity results}

\item{ncomp}{which component make the map for}

\item{main}{main title for the plot}

\item{col}{a color map for the image}

\item{xaxt}{x-axis type}

\item{yaxt}{y-axis type}

\item{...}{other image parameters}
}
\description{
Make an image with map of resolved concentrations
}
\details{
The method will work only if original spectral data
has attributes \code{width} and \code{height} pointing out
that it is a hypercube
}

