% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mcrals.R
\name{mcrals}
\alias{mcrals}
\title{Angle constrained MCR-ALS.}
\usage{
mcrals(D, ncomp, Ct = NULL, St = NULL, do.angle = T, f = 0.05,
  wavelength = NULL, max.niter = 500)
}
\arguments{
\item{D}{a matrix with spectral values}

\item{ncomp}{number of pure components to identify}

\item{Ct}{initial estimates for pure contributions}

\item{St}{initial estimates for pure spectra}

\item{do.angle}{logical, should the angle constraint be used}

\item{f}{how much of mean should be added for angle constraint}

\item{wavelength}{a vector with wavelength if necessary}

\item{max.niter}{maximum number of iterations}
}
\value{
The method returns results of unmixing as an object of \code{mcrals} class with following fields.
\item{spec }{resolved spectra of pure components.}
\item{conc }{resolved concentrations of pure components.}
\item{ncomp }{number of resolved components.}
\item{resvar }{residual variance for each component.}
}
\description{
Angle constrained MCR-ALS.
}
\details{
The method implements a simple MCR-ALS algorithm with non-negativity constraints applied both to
spectra and contributions by setting all negative values to zero. In addition an angle constraint
can be applied by adding a small portion (default 5%) of mean to spectra or contributions [1].

If initial estimates for contributions and spectra are not provided the estimates will
be generated with random numbers.
}
\examples{
# unmix hyperspectral data with and without angle constraint and compare the
# results for first two components
data(puredata)

res1 = mcrals(hsi$D, 3, do.angle = FALSE, wavelength = hsi$wavelength)
res2 = mcrals(hsi$D, 3, do.angle = TRUE, wavelength = hsi$wavelength)

summary(res1)
summary(res2)

# because initial estimates are made with random numbers
# the order of components in the results is different
par(mfrow = c(2, 1))
plotSpectra(res1)
plotSpectra(res2)
par(mfrow = c(1, 1))

par(mfrow = c(3, 2))
for (i in 1:3)
{
 image(res1, i)
 image(res2, i)
}
par(mfrow = c(1, 1))
}
\references{
1. W. Windig, M.R. Keenan, Appl. Spectrosc. 65 (2011) 349-357.
}
\seealso{
The class has numerous methods for investigating and refininf the unmixing results.
\tabular{ll}{
 \code{print} \tab prints information about a \code{purity} object.\cr
 \code{summary} \tab shows summary statistics for the object.\cr
 \code{plot} \tab plot overview of unmixing results.\cr
 \code{\link{plotSpectra}} \tab shows a plot with resolved spectra.\cr
 \code{\link{plotConcentrations}} \tab shows a plot with resolved concentrations.\cr
 \code{\link{plotVariance}} \tab shows a plot with residual variance vs. number of components.\cr
 \code{\link{plotSpectraComparison}} \tab compares resolved and ethalon spectra if available.\cr
 \code{\link{image.mcrals}} \tab if data is a hypercube, shows resolved concetration map as an image.\cr
}
}

