% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mcrals.R
\name{plot.mcrals}
\alias{plot.mcrals}
\title{Overview plot for mcrals object}
\usage{
\method{plot}{mcrals}(x, ...)
}
\arguments{
\item{x}{a  (object of class \code{mcrals})}

\item{...}{other arguments}
}
\description{
Shows a set of plots for MCR-ALS results.
}
\details{
See examples in help for \code{\link{mcrals}} function.
}

