% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/purity.R
\name{plot.purity}
\alias{plot.purity}
\title{Overview plot for purity object}
\usage{
\method{plot}{purity}(x, comp = 1:x$ncomp, ...)
}
\arguments{
\item{x}{a  (object of class \code{purity})}

\item{comp}{for which component(s) make the plot for}

\item{...}{other arguments}
}
\description{
Shows a set of plots () for .
}
\details{
See examples in help for \code{\link{purity}} function.
}

