% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{plotConcentrations}
\alias{plotConcentrations}
\title{Concentrations plot}
\usage{
plotConcentrations(obj, comp = 1:obj$ncomp, origspec = NULL,
  col = getColors(length(comp)), lty = 1, lwd = 1, xlab = "Mixings",
  ylab = "", main = "Resolved concentrations", ...)
}
\arguments{
\item{obj}{object with unmixing results}

\item{comp}{which components make the plot for}

\item{origspec}{original spectral data used for unmixing}

\item{col}{color for the plot lines}

\item{lty}{line type for plot lines}

\item{lwd}{line width for plot lines}

\item{xlab}{label for x axis}

\item{ylab}{label for y axis}

\item{main}{main title for the plot}

\item{...}{other parameters for matplot function}
}
\description{
Shows a plot with resolved concentration profiles
}

