% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{plotPurity}
\alias{plotPurity}
\title{Purity spectra plot}
\usage{
plotPurity(obj, ncomp = 1, origspec = NULL, col = c("#62B7EF", "#D84E5E",
  "#F0F0F0"), lty = 1, lwd = 1, xlab = "Wavenumbers",
  ylab = c("Weighted std", "Purity"), main = sprintf("Purity for C\%d",
  ncomp))
}
\arguments{
\item{obj}{object with unmixing results}

\item{ncomp}{which component to make the plot for}

\item{origspec}{original spectral data used for unmixing}

\item{col}{colors for the std, purity and original spectra}

\item{lty}{line type for the component spectra}

\item{lwd}{line width for the component spectra}

\item{xlab}{label for x axis}

\item{ylab}{label for y axis}

\item{main}{main title for the plot}
}
\description{
Shows a plot with purity spectra for each or user selected component
}

