% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/supure.R
\name{plotResiduals}
\alias{plotResiduals}
\title{Residuals plot for supure object}
\usage{
plotResiduals(obj, type = "b", col = getColors(1), pch = 16,
  xlab = "Number of components", ylab = "Variance, \%",
  main = "Residual variance", ...)
}
\arguments{
\item{obj}{object with supure results}

\item{xlab}{label for x axis}

\item{ylab}{label for y axis}

\item{main}{main title for the plot}
}
\description{
Residuals plot for supure object
}

