% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{plotSpectra}
\alias{plotSpectra}
\title{Spectral plot}
\usage{
plotSpectra(obj, comp = 1:obj$ncomp, origspec = NULL, show.labels = T,
  show.legend = T, legend.str = NULL, ylim = NULL, xlim = NULL,
  col = getColors(length(comp)), lty = 1, lwd = 1, xlab = "Wavenumbers",
  ylab = "Intensity", main = "Resolved spectra", ...)
}
\arguments{
\item{obj}{object with unmixing results}

\item{comp}{which components make the plot for}

\item{origspec}{original spectral data used for unmixing}

\item{show.labels}{logical, show or not labels for the selected pure variables on the plot}

\item{show.legend}{logical, show or not a legend for the plot}

\item{legend.str}{a vector with legend items}

\item{ylim}{limits for y-axis}

\item{xlim}{limits for x-axis}

\item{col}{color for the component spectra}

\item{lty}{line type for the component spectra}

\item{lwd}{line width for the component spectra}

\item{xlab}{label for x axis}

\item{ylab}{label for y axis}

\item{main}{main title for the plot}

\item{...}{other maplot parameters}
}
\description{
Shows a plot with resolved spectra
}

