% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{plotSpectraComparison}
\alias{plotSpectraComparison}
\title{Compares resolved and original pure component spectra}
\usage{
plotSpectraComparison(obj, ospectra, ncomp = 1, col = getColors(2),
  lwd = c(2, 1), lty = c(1, 1), xlab = "Wavenumbers",
  ylab = "Intensity", main = NULL, show.stat = F, ...)
}
\description{
Compares resolved and original pure component spectra
}

