% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{plotVariance}
\alias{plotVariance}
\title{Residual variance plot for results of unmixing}
\usage{
plotVariance(obj, type = "b", col = getColors(1), pch = 16,
  xlab = "Number of components", ylab = "Variance, \%",
  main = "Residual variance", ...)
}
\arguments{
\item{obj}{object with unmixing results}

\item{xlab}{label for x axis}

\item{ylab}{label for y axis}

\item{main}{main title for the plot}
}
\description{
Residual variance plot for results of unmixing
}

