% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mcrals.R
\name{print.mcrals}
\alias{print.mcrals}
\title{Print method for mcrals object}
\usage{
\method{print}{mcrals}(x, ...)
}
\arguments{
\item{x}{unmixing results (object of class \code{mcrals})}

\item{...}{other arguments}
}
\description{
Prints information about the object structure
}

