% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/purity.R
\name{print.purity}
\alias{print.purity}
\title{Print method for purity object}
\usage{
\method{print}{purity}(x, ...)
}
\arguments{
\item{x}{unmixing results (object of class \code{purity})}

\item{...}{other arguments}
}
\description{
Prints information about the object structure
}

