% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/purity.R
\name{purity}
\alias{purity}
\title{Spectral unmixing with pure variables}
\usage{
purity(spectra, ncomp, offset = 5, by.spec = T, use.deriv = 0,
  savgol = c(2, 5, 2), exclude = NULL, wavelength = NULL)
}
\arguments{
\item{ncomp}{number of pure components to identify}

\item{offset}{offset in percent for calculation of parameter alpha}

\item{by.spec}{logical, should the algorithm works with spectra (rows) or with concentrations (columns)}

\item{use.deriv}{shall a second derivative be used (0 - no, 1 - only for unmixing, 2 - both for locating pure
variables and for unmixing)}

\item{savgol}{vector of parameters for Savitzky-Goley filter: derivative order, width of filter and polynomial order}

\item{exclude}{vector of column numbers that should be excluded when pure variables are being detected}

\item{wavelength}{a vector with wavelength if necessary}

\item{data}{a matrix with spectral data}
}
\value{
The method returns an object of \code{purity} class - a list with with following fields:
\item{spec }{resolved spectra of pure components.}
\item{conc }{resolved concentrations of pure components.}
\item{ncomp }{number of resolved components.}
\item{purvars }{vector with indices for pure variables.}
\item{purityspec }{matrix with purity values for each resolved components.}
\item{stdspec }{matrix with weighted standard deviation values.}
\item{purity }{vector with purity values for resolved components.}
\item{wavelength }{vector with wavelength/wavebands/wavenumbers.}
}
\description{
\code{purity} is a class for resolving pure spectra and contributions from spectra of mixtures
using pure variable approach.
}
\details{
The unmixing is done with two steps. First, pure variables are identified using methods, described
in [1-2]. Then the selected pure variables are used to resolve the pure spectra and contributions
with ordinary least squares.

By default method works with spectra, rows of the data matrix, but it is possible to make the
unmixing using concentration profiles (with parameter \code{by.spec} set to \code{FALSE}). Also
an inverse second derivative of the data can be used for either unmixing step or for both
identifying the pure variables and for unmixing in case if spectral peaks are overlapped.
}
\examples{
## 1. Resolving of Raman spectra of carbonhydrates

# get the data
data(puredata)

# do unmixing with standard settings
res = purity(carbs$D, 3, wavelength = carbs$wavelength)
summary(res)
plot(res)

# compare the unmixed spectra with ethalon
par(mfrow = c(3, 1))
plotSpectraComparison(res, carbs$S, 1)
plotSpectraComparison(res, carbs$S, 2)
plotSpectraComparison(res, carbs$S, 3)
par(mfrow = c(1, 1))

# do unmixing with two different settings and compare purity for first component
res1 = purity(carbs$D, 3, offset = 5, use.deriv = F)
res2 = purity(carbs$D, 3, offset = 15, use.deriv = T, savgol = c(2, 5, 2))
par(mfrow = c(2, 1))
plotPurity(res1, 1)
plotPurity(res2, 1)
par(mfrow = c(1, 1))

## 2. Hyperspectral image of oil-in-water emulsion

# get the data
data(puredata)

# it is important that variable has proper attributes
show(attr(hsi$D, 'width'))
show(attr(hsi$D, 'height'))

# do unmixing
res = purity(hsi$D, 3, offset = 3, wavelength = hsi$wavelength)
summary(res)
plot(res)

# show concentration maps
par(mfrow = c(1, 3))
image(res, 1)
image(res, 2)
image(res, 3)
}
\references{
1. W. Windig, J. Guilment, Anal. Chem. 63 (1991) 1425–1432.
2. W. Windig, Chemom. Intell. Lab. Syst. 23 (1994) 71–86.
}
\seealso{
The class has numerous methods for exploring and refining the unmixing results.
\tabular{ll}{
 \code{print} \tab prints information about a \code{purity} object.\cr
 \code{summary} \tab shows summary statistics for the object.\cr
 \code{plot} \tab plot overview of unmixing results.\cr
 \code{\link{plotPurity}} \tab shows purity values and weighted standard deviation as a line plot.\cr
 \code{\link{plotSpectra}} \tab shows a plot with resolved spectra.\cr
 \code{\link{plotConcentrations}} \tab shows a plot with resolved concentrations.\cr
 \code{\link{plotVariance}} \tab shows a plot with residual variance vs. number of components.\cr
 \code{\link{plotSpectraComparison}} \tab compares resolved and ethalon spectra if available.\cr
 \code{\link{image.purity}} \tab if data is a hypercube, shows resolved concetration map as an image.\cr
 \code{\link{explore}} \tab a GUI tool to explore and tune the results (requires `shiny`).\cr
}
The steps of the algorithms can be also applied separately with functions
\code{\link{getpurevars}} and \code{\link{unmix}}.
}

