% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/purity.R
\name{savgol}
\alias{savgol}
\title{Savytzky-Golay filter}
\usage{
savgol(data, params)
}
\arguments{
\item{data}{a matrix with data values}

\item{params}{vector with parameters: derivative order, width of filter and polynomial order}
}
\description{
Applies Savytzky-Golay filter to the rows of data matrix
}

