% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mcrals.R
\name{summary.mcrals}
\alias{summary.mcrals}
\title{Summary method for mcrals object}
\usage{
\method{summary}{mcrals}(object, ...)
}
\arguments{
\item{object}{unmixing results (object of class \code{mcrals})}

\item{...}{other arguments}
}
\description{
Shows summary statistics for mcrals objects.
}

