% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/purity.R
\name{summary.purity}
\alias{summary.purity}
\title{Summary method for purity object}
\usage{
\method{summary}{purity}(object, ...)
}
\arguments{
\item{object}{unmixing results (object of class \code{purity})}

\item{...}{other arguments}
}
\description{
Shows summary statistics for purity objects.
}

