% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/purity.R
\name{unmix}
\alias{unmix}
\title{Unmix spectral data with least squares method}
\usage{
unmix(D, Dr, by.spec = T)
}
\arguments{
\item{D}{matrix with the spectral data}

\item{Dr}{a matrix with concentration estimates (e.g. columns of D with pure variables)}

\item{by.spec}{logical, should the algorithm works with spectra (rows) or with concentrations (columns)}
}
\value{
Returns a list with two matrices, `conc` for concentrations and `spec` for spectra as well
as a vector with residual variance values
}
\description{
\code{unmix} decomposes the spectral data to a matrix with concentrations and
a matrix with spectra of pure components using set of pure variables.
}

