% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/supure.R
\name{plotSpectra}
\alias{plotSpectra}
\title{Spectral plot}
\usage{
plotSpectra(obj, comp = 1:obj$ncomp, origspec = NULL, show.labels = T,
  col = getColors(length(comp)), lty = 1, lwd = 1, xlab = "Wavenumbers",
  ylab = "Intensity", main = "Resolved spectra", ...)
}
\arguments{
\item{obj}{object with supure results}

\item{comp}{which components make the plot for}

\item{show.labels}{logical, show or not labels for the selected pure variables on the plot}

\item{col}{color for the component spectra}

\item{lty}{line type for the component spectra}

\item{lwd}{line width for the component spectra}

\item{xlab}{label for x axis}

\item{ylab}{label for y axis}

\item{main}{main title for the plot}
}
\description{
Shows a plot with original spectral data, evaluated spectra of pure components and selected
pure variables
}

