% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/supure.R
\name{savgol}
\alias{savgol}
\title{Savytzky-Golay filter}
\usage{
savgol(data, params)
}
\arguments{
\item{data}{a matrix with data values}

\item{width}{width of the filter window}

\item{porder}{order of polynomial used for smoothing}

\item{dorder}{order of derivative to take (0 - no derivative)}
}
\description{
Applies Savytzky-Golay filter to the rows of data matrix
}

