% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/supure.R
\name{supure}
\alias{supure}
\title{Makes unmixing of set of spectra to pure components.}
\usage{
supure(spectra, ncomp, offset = 5, use.deriv = 0, savgol = c(2, 5, 2),
  exclude = NULL, wavelength = NULL)
}
\arguments{
\item{ncomp}{number of pure components to identify (optional)}

\item{use.deriv}{shall a second derivative be used (0 - no, 1 - only for unmixing, 2 - both for locating pure
variables and for unmixing)}

\item{savgol}{parameters for Savitzky-Goley filter: derivative order, width of filter and polynomial order}

\item{exclude}{vector of column numbers that should be excluded when pure variables are being detected}

\item{wavelength}{a vector with wavelength if necessary}

\item{data}{a matrix with spectral values}
}
\value{
The method returns an object of \code{supure} class with following fields.
\code{supure } object - list with several fields, including:
\item{spec }{resolved spectra of pure components.}
\item{conc }{resolved concentrations of pure components.}
\item{purvars }{vector with indices for pure variables.}
\item{purity }{vector with values for purity function.}
}
\description{
Makes unmixing of set of spectra to pure components.
}

