% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/supure.R
\name{supure}
\alias{supure}
\title{Makes unmixing of set of spectra to pure components.}
\usage{
supure(spectra, ncomp, suggest = NULL, savgol.deriv = 2,
  savgol.width = NULL, savgol.porder = 2, do.norm = TRUE,
  wavelength = NULL)
}
\arguments{
\item{ncomp}{number of pure components to identify (optional)}

\item{suggest}{a sequence with variable numbers suggested as pure variables for each component}

\item{savgol.deriv}{a derivative order (if 0 no derivative will be used)}

\item{savgol.width}{a width of filter used to smooth signal prior to take a derivative}

\item{savgol.porder}{a polynomial order used for smoothing}

\item{do.norm}{logical, normalize the spectral data to unit area prior to unmixing or not}

\item{wavelength}{a vector with wavelength if necessary}

\item{data}{a matrix with spectral values}
}
\value{
A \code{supure} object - list with several fields, including:
 \item{spec}{resolved spectra of pure components.}
 \item{conc}{resolved concentrations of pure components.}
}
\description{
Makes unmixing of set of spectra to pure components.
}

