% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/supure.R
\name{unmix}
\alias{unmix}
\title{Unmix spectral data with least squares method}
\usage{
unmix(D, purevars, savgol = c(0, 1, 1))
}
\arguments{
\item{D}{matrix with the spectral data}

\item{purevars}{a vector with pure variables}

\item{savgol}{parameters for Savitzky-Goley filter: derivative order, width of filter and polynomial order}
}
\value{
Returns a list with two matrices, `conc` for concentrations and `spec` for spectra
}
\description{
`unmix` decomposes the spectral data to a matrix with concentrations and
a matrix with spectra of pure components using set of pure variables.
}

